requirements.txt
discord.py
openai
python-dotenv
watchdog
requests
alive-progress
weaviate-client
